pyproject.toml
[project]
name = "fastapi-notes"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "celery[redis,timezone]>=5.5.3",
    "fastapi>=0.115.14",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pyjwt==2.8.0",
    "python-decouple==3.8",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "pytz>=2025.2",
    "redis>=5.2.1",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.41",
    "uvicorn[standard]>=0.35.0",
]
